+++ packages/react-server/src/index.ts
@@ -1,4 +1,3 @@
-// release comment
 export {createServer} from './server';
 export {createRender} from './render';
 export type {Context} from './render';